Good to go until shift table selection is ready

+++ models/job.js
@@ -4,7 +4,7 @@ module.exports = function(sequelize, DataTypes) {
 		user_id: DataTypes.INTEGER,
 		startDate: DataTypes.DATE,
 		endDate: DataTypes.DATE,
-		wage: DataTypes.DECIMAL,
+		wage: DataTypes.FLOAT(10,2),
 		isReal: DataTypes.BOOLEAN,
 		userJobMenu: DataTypes.INTEGER,
 		comments: DataTypes.TEXT
